+++ cryopid/trunk/src/cryopid.h
@@ -33,7 +33,7 @@
 void write_stub(int fd, long offset);
 
 /* common.c */
-int syscall_check(int retval, int can_be_fake, char* desc, ...);
+long syscall_check(int retval, int can_be_fake, char* desc, ...);
 void safe_read(int fd, void* dest, size_t count, char* desc);
 void *xmalloc(int len);
 unsigned int checksum(char *ptr, int len, unsigned int start);
